Restructured back into each cpp
Started implementing gestures for android
For now only line motion and no midpoints because I can't figure it out.

have to add other types of input (tap, key,...) but key will conform to the keyboard and modify the state accordingly.

This way the android keyboard will not be mapped appropriately to standard ASCII keyboard, will have to either hard code or just modify what the keys really do or find a way to modify string without really getting into keys directly (which should also be done for desktop keyboard).
